Fix the rest of the class loading issues & add more helper methods

+++ api/src/main/java/mods/thecomputerizer/theimpossiblelibrary/api/client/gui/widget/Scrollable.java
@@ -2,6 +2,6 @@ package mods.thecomputerizer.theimpossiblelibrary.api.client.gui.widget;
 
 public interface Scrollable {
     
-    boolean scrollDown(int scroll);
-    boolean scrollUp(int scroll);
+    boolean scrollDown(double scroll);
+    boolean scrollUp(double scroll);
 }
